refactor: split test_util to separate modules
