+++ Cargo.toml
@@ -27,7 +27,7 @@ rayon = "1.5.2"
 static_assertions = "1.1.0"
 strum = {version = "0.24.0", optional = true}
 xias = "0.3.0"
-derive-trait = "0.0.2"
+derive-trait = "0.0.3"
 
 [features]
 default = ["debug-entity-rc"]
